+++ app/controllers/sources_controller.rb
@@ -218,6 +218,7 @@ class SourcesController < ApplicationController
     params[:project_id] = sessions_current_project_id
     params.permit(
       :author,
+      :author_ids_or,
       :citations,
       :documents,
       :exact_author,