+++ app/javascript/vue/tasks/sources/filter/components/filters/authors.vue
@@ -23,6 +23,12 @@
           roles: ['SourceAuthor', 'SourceEditor']
         }"
         @selected="addAuthor"/>
+      <label>
+        <input
+          v-model="source.author_ids_or"
+          type="checkbox">
+        Any
+      </label>
     </fieldset>
     <display-list
       :list="authors"
@@ -84,6 +90,7 @@ export default {
     const params = URLParamsToJSON(location.href)
     this.source.author = params.author
     this.source.exact_author = params.exact_author
+    this.source.author_ids_or = params.author_ids_or
     if (params.author_ids) {
       params.author_ids.forEach(id => {
         GetPeople(id).then(response => {